Move mock value to utils file
