+++ packages/client/src/tests/util.tsx
@@ -2294,6 +2294,175 @@ export function createRouterProps<
   return { location, history, match }
 }
 
+export const mockRoles = {
+  data: {
+    getSystemRoles: [
+      {
+        value: 'FIELD_AGENT',
+        roles: [
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'Healthcare Worker'
+              },
+              {
+                lang: 'fr',
+                label: 'Professionnel de Santé'
+              }
+            ]
+          },
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'Police Officer'
+              },
+              {
+                lang: 'fr',
+                label: 'Agent de Police'
+              }
+            ]
+          },
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'Social Worker'
+              },
+              {
+                lang: 'fr',
+                label: 'Travailleur Social'
+              }
+            ]
+          },
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'Local Leader'
+              },
+              {
+                lang: 'fr',
+                label: 'Leader Local'
+              }
+            ]
+          }
+        ],
+        active: true
+      },
+      {
+        value: 'REGISTRATION_AGENT',
+        roles: [
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'Registration Agent'
+              },
+              {
+                lang: 'fr',
+                label: "Agent d'enregistrement"
+              }
+            ]
+          }
+        ],
+        active: true
+      },
+      {
+        value: 'LOCAL_REGISTRAR',
+        roles: [
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'Local Registrar'
+              },
+              {
+                lang: 'fr',
+                label: 'Registraire local'
+              }
+            ]
+          }
+        ],
+        active: true
+      },
+      {
+        value: 'LOCAL_SYSTEM_ADMIN',
+        roles: [
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'Local System_admin'
+              },
+              {
+                lang: 'fr',
+                label: 'Administrateur système local'
+              }
+            ]
+          }
+        ],
+        active: true
+      },
+      {
+        value: 'NATIONAL_SYSTEM_ADMIN',
+        roles: [
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'National System_admin'
+              },
+              {
+                lang: 'fr',
+                label: 'Administrateur système national'
+              }
+            ]
+          }
+        ],
+        active: true
+      },
+      {
+        value: 'PERFORMANCE_MANAGEMENT',
+        roles: [
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'Performance Management'
+              },
+              {
+                lang: 'fr',
+                label: 'Gestion des performances'
+              }
+            ]
+          }
+        ],
+        active: true
+      },
+      {
+        value: 'NATIONAL_REGISTRAR',
+        roles: [
+          {
+            labels: [
+              {
+                lang: 'en',
+                label: 'National Registrar'
+              },
+              {
+                lang: 'fr',
+                label: 'Registraire national'
+              }
+            ]
+          }
+        ],
+        active: true
+      }
+    ]
+  }
+}
+
 export const mockFetchRoleGraphqlOperation = {
   request: {
     query: getSystemRolesQuery,
