+++ packages/client/src/views/SysAdmin/Team/user/userCreation/SignatureForm.test.tsx
@@ -20,6 +20,7 @@ import {
   mockOfflineDataDispatch,
   mockFetchRoleGraphqlOperation,
   mockDataWithRegistarRoleSelected,
+  mockRoles,
   mockUserGraphqlOperation
 } from '@client/tests/util'
 import { waitForElement } from '@client/tests/wait-for-element'
@@ -29,13 +30,13 @@ import { ReactWrapper } from 'enzyme'
 import * as React from 'react'
 import { roleQueries } from '@client/forms/user/query/queries'
 import { Mock, describe, expect } from 'vitest'
-import { mockRoles } from './CreateNewUser.test'
 
 describe('signature upload tests', () => {
   const { store, history } = createStore()
   let testComponent: ReactWrapper
 
   beforeEach(async () => {
+    ;(roleQueries.fetchRoles as Mock).mockReturnValue(mockRoles)
     await store.dispatch(offlineDataReady(mockOfflineDataDispatch))
   })
